backend/requirements.txt
# ─── RDVPro Backend Dependencies ─────────────────────────────────────────

# Web framework
flask==3.0.2
flask-cors==4.0.0

# Auth
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1

# Database
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-migrate==4.0.7

# Validation
marshmallow==3.21.0
flask-marshmallow==1.2.0

# Email
flask-mail==0.10.0

# Task scheduling (for reminders)
apscheduler==3.10.4

# AI recommendations (scikit-learn based)
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1

# Utilities
python-dotenv==1.0.1
gunicorn==21.2.0

# Testing
pytest==8.1.0
pytest-flask==1.3.0
factory-boy==3.3.0
